Feat: 결제 알림 기능 Event Listener 처리 및 GlobalException을 통한 확장성 관리

+++ demo/src/main/java/com/example/demo/mypage/carItem/controller/CartController.java
@@ -1,6 +1,8 @@
 package com.example.demo.mypage.carItem.controller;
 
 import com.example.demo.login.global.annotation.Member;
+import com.example.demo.login.member.exception.exceptions.MemberErrorCode;
+import com.example.demo.login.member.exception.exceptions.MemberException;
 import com.example.demo.mypage.carItem.controller.dto.*;
 import com.example.demo.mypage.carItem.service.CartService;
 import lombok.RequiredArgsConstructor;
@@ -17,6 +19,9 @@ public class CartController {
     @PostMapping("/add")
     public ResponseEntity<Void> addToCart(@RequestBody AddCartRequest request,
                                           @Member Long memberId) {
+        if (memberId == null) {
+            throw new MemberException(MemberErrorCode.NOT_AUTHENTICATED);
+        }
         cartService.addToCart(request.getProductDetailId(), request.getQuantity(), memberId);
         return ResponseEntity.ok().build();
     }
@@ -24,6 +29,9 @@ public class CartController {
     @PutMapping("/update")
     public ResponseEntity<Void> updateQuantity(@RequestBody UpdateQuantityRequest request,
                                                @Member Long memberId) {
+        if (memberId == null) {
+            throw new MemberException(MemberErrorCode.NOT_AUTHENTICATED);
+        }
         cartService.updateQuantityByCartItemId(
                 memberId,
                 request.getCartItemId(),
@@ -34,6 +42,9 @@ public class CartController {
 
     @DeleteMapping("/clear")
     public ResponseEntity<Void> clearCart(@Member Long memberId) {
+        if (memberId == null) {
+            throw new MemberException(MemberErrorCode.NOT_AUTHENTICATED);
+        }
         cartService.clearCart(memberId);
         return ResponseEntity.ok().build();
     }
@@ -41,12 +52,18 @@ public class CartController {
     @DeleteMapping("/delete")
     public ResponseEntity<Void> deleteSelectedItems(@Member Long memberId,
                                                     @RequestBody CartItemDeleteRequest request) {
+        if (memberId == null) {
+            throw new MemberException(MemberErrorCode.NOT_AUTHENTICATED);
+        }
         cartService.deleteSelectedItems(memberId, request.getCartItemIds());
         return ResponseEntity.ok().build();
     }
 
     @GetMapping
     public ResponseEntity<CartSummaryResponse> getCart(@Member Long memberId) {
+        if (memberId == null) {
+            throw new MemberException(MemberErrorCode.NOT_AUTHENTICATED);
+        }
         return ResponseEntity.ok(cartService.getCart(memberId));
     }
 }
